+++ .github/workflows/deploy-dev.yml
@@ -1,4 +1,4 @@
-name: Android CI
+name: Build and Deploy to Firebase Hosting on dev
 
 on:
   push:
